This commit has a large rename of files. Let me explain-
When RH.org fired up, it wasn't sorting the filenames.
Now, that gives us a problem.. It was relying on the ext order of the bits.
Oops. Sorry, I didn't think about that, I released in a hurry.

Anyway, to keep compatibility, I've ensured that the fileorder from RH.org
is now reflected in the files themselves. This was done by naming each one with up to
two sets of numbers

0001#01Body/01.png

What the 0001# before the pound sign means does is ensure the
code is sorted with that gif first, when assigning them.

Anyway, this makes the code messier, but it will be the same everywhere now.
I'll clean up the filenames later, if anyone cares.
Or, you're welcome to!

Renaming them to drop the numbers in the basefilename should be perfectly OK.
Try it ;)
-CPD